backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
groq
sentence-transformers
faiss-cpu
rank_bm25
langdetect
deep-translator
PyPDF2
python-docx
numpy
scikit-learn
# For real predictor (uncomment when model URL is ready):
huggingface_hub
xgboost
joblib
pandas
